fix migration issue

+++ db/migrate/20250107223747_remove_uuid_from_shared_watchlists.rb
@@ -2,6 +2,6 @@
 
 class RemoveUuidFromSharedWatchlists < ActiveRecord::Migration[7.2]
   def change
-    remove_column :shared_watchlists, :uuid, :string
+    remove_column :shared_watchlists, :uuid, :string if column_exists?(:shared_watchlists, :uuid)
   end
 end
